requirements.txt
click==8.1.3
cycler==0.10.0
Flask==2.2.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.3.1
MarkupSafe==2.1.1
matplotlib==3.4.2
numpy==1.21.1
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
slack-bolt==1.14.3
slack-sdk==3.18.1
Werkzeug==2.2.1
zipp==3.8.1
python-dotenv~=0.20.0